+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 ### Changes
 
+* `NonNilCheck` offense reporting and autocorrect are configurable to include semantic changes. ([@hannestyden][])
 * The parameters `AllCops/Excludes` and `AllCops/Includes` with final `s` only give a warning and don't halt `rubocop` execution. ([@jonas054][])
 * The `GuardClause` cop is no longer irgnoring a one-line body by default - see configuration. ([@geniou][])